bump middleware (#1783)
* bump middleware

* workaround bad numeric widening warnings on 2.10

+++ server/src/main/scala/org/ensime/server/DocsHandler.scala
@@ -97,7 +97,7 @@ class DocsHandler(docs: DocJarReading) extends SimpleChannelInboundHandler[FullH
         }
         res
     }
-    HttpUtil.setContentLength(res, res.content().readableBytes())
+    HttpUtil.setContentLength(res, res.content().readableBytes().toLong)
 
     // Send the response and close the connection if necessary.
     val f: ChannelFuture = ctx.channel().writeAndFlush(res)
